+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## [Unreleased]
+## 2025-10-29 - [0.3.5]
 
 ### Bug fixes
 
@@ -207,7 +207,7 @@
 
 First beta release
 
-[unreleased]: https://github.com/mondeja/hledger-fmt/compare/v0.3.4...master
+[0.3.5]: https://github.com/mondeja/hledger-fmt/compare/v0.3.4...v0.3.5
 [0.3.4]: https://github.com/mondeja/hledger-fmt/compare/v0.3.3...v0.3.4
 [0.3.3]: https://github.com/mondeja/hledger-fmt/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/mondeja/hledger-fmt/compare/v0.3.1...v0.3.2